DSP process now uses blocking writes by default
When the program wishes to write data to the hardware buffers, there often needs
to be a guarantee that the data has actually been recieved and uploaded
properly.  By default, hardware reads will continue to be non-blocking (e.g. the
child process will read the data and write it to shared memory, but will not
wait until the parent process obtains the data from shared memory.  In contrast,
when the parent process writes data to shared memory, it will wait until the
child process has retrieved the data from shared memory and uploaded it to the
hardware buffer.  Then, and only then, will the buffer.write() or buffer.set()
method return in the parent process.  To default to the original non-blocking
behavior, use buffer.write(data, timeout=0) or buffer.set(data, timeout=0).