Added device-specific utilities for the RZ6
The RZ6 audio-out macro is very CPU-intensive at high sampling rates.  Much of
the computations done via the macro on *every* cycle can also be done in the
software and uploaded to the circuit.  I've finished working out all the details
of the computation for 1) computing the integer value that has to be poked into
the RZ6 memory to set the hardware attenuators and 2) converting the waveform
voltage samples into an array of integers that can be uploaded to an integer
buffer in the circuit which then "pokes" the values into the appropriate
location in device memory.  At a sampling rate of 100 kHz, this reduces the CPU
overhead of using the AudioOut macro by a factor of 14%.

Eventually I would like to add device-specific utility functions under the
tdt.device namespace (e.g. tdt.device.RZ6).  This reflects a start in this
direction.