Bugfix to test code for RZ6
When using DSPBuffer.acquire, we need to make sure that the block_size of the
DSPBuffer is consistent with the expected number of samples to be acquired.  If
block_size is not a divisor of the samples to be acquired, we may not pull in
all the data.  For now, the simplest approach is to set block_size to 1.