Modified cset_tag to return the converted value
A common use case is to use cset_tag to convert the value and upload it to the
RPvds circuit.  For exmaple, if you wish to set the trial duration tag in the
buffer to 6 seconds, you would have cset_tag convert the value for you.
However, you may also wish to know the actual value the tag was set to.
DSPCircuit.cset_tag now returns this value.