Working draft of a network-aware of the TDTPy library
This is a fully functional draft that implements a remote procedure call (RPC)
client/server architecture.  Since the TDT ActiveX library is not thread-safe,
there were concurrency issues when breaking up the code into separate processes
(e.g. an acquisition process could be separate from the stimulus generation
process).  To implement this, we decided to go the socket route where a single
RPC server handles all client requests in a sequential fashion (thus handling
concurrent access issues from multiple processes).

This approach has the additional benefit of allowing one to run the client
software remotely on another computer (on *any* operating system, not just
Windows).  However, the server must be a Windows computer since the server-side
code is essentially a wrapper around the ActiveX RPcoX and zBUS objects.

As far as I can tell, the speed seems reasonable; however, there are likely
additional speed enhancements that can be introduced as we flesh out this
library.  Right now, it's (unfortunately) no faster than the non-socket
approach; however, this is due (in part) to the lack of non-blocking IO on the
client side.  Furthermore, the timeout of 10 seconds on the server side likely
should be replaced with something more intelligent.