Updated TDTPy abstraction layer to allow an interface argument
TDT currently has two interfaces, the USB and GB interface.  We have only tested
the GB interface in our lab; however, other labs may require the ability to
connect via the USB interface.  To allow this, I've updated the connect_*
utility functions as well as the DSPCircuit/DSPProject class with the
appropriate keyword argument, 'interface'.  Note that it still defaults to
'GB', so existing code does not have to be changed.