Updated TDTPy documentation
Added a note that numpydoc is required for building the docs.

Now that TDTPy is listed on PyPI, it's no longer necessary to provide
instructions on how to install from the source-code repository. Most people who
want to use TDTPy will most likely not make changes.

Eventually I may add a "for developers" section explaining how contributions can be made to the project.