Add lock for thread-safety
Even after adding the latch trigger in the circuit, we have to make
sure that we read the index/cycle tags concurrently. This fixes a bug
where I had two separate threads (an analog output and analog input) that
were competing with each other to use the same latch trigger and check
the cycle/index tags at the same time.